requirements.txt
more-itertools
Pillow